# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DropdownOption {
  id: number;
  category: string;
  value: string;
  sortOrder: number;
  createdAt: string;
}

export interface DropdownData {
  payroll: string[];
  company: string[];
  work: string[];
  pending: string[];
  officer: string[];
  assigned: string[];
}

export interface AddDropdownOptionBody {
  value: string;
}

export interface DeleteDropdownOptionBody {
  value: string;
}

export interface Submission {
  id: number;
  payroll: string;
  company: string;
  work: string;
  detail1: string;
  /** @nullable */
  detail2: string | null;
  pending: string;
  /** @nullable */
  remarks: string | null;
  /** @nullable */
  officer: string | null;
  /** @nullable */
  assigned: string | null;
  submittedAt: string;
}

export interface SubmissionList {
  items: Submission[];
  total: number;
}

export interface CreateSubmissionBody {
  payroll: string;
  company: string;
  work: string;
  detail1: string;
  /** @nullable */
  detail2?: string | null;
  pending: string;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  officer?: string | null;
  /** @nullable */
  assigned?: string | null;
}

export interface WorkCount {
  work: string;
  count: number;
}

export interface PendingCount {
  pending: string;
  count: number;
}

export interface OfficerCount {
  officer: string;
  total: number;
  done: number;
  pending: number;
}

export interface DoerCount {
  doer: string;
  total: number;
  done: number;
  pending: number;
}

export interface SubmissionSummary {
  total: number;
  todayCount: number;
  byWork: WorkCount[];
  byPending: PendingCount[];
  byOfficer: OfficerCount[];
  byDoer: DoerCount[];
  recentSubmissions: Submission[];
}

export type ListSubmissionsParams = {
  limit?: number;
  offset?: number;
};

export type GetSubmissionSummaryParams = {
  /**
   * Filter from this date (ISO format YYYY-MM-DD)
   */
  startDate?: string;
  /**
   * Filter up to this date (ISO format YYYY-MM-DD)
   */
  endDate?: string;
};