# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all dropdown options for the form (payroll, company, work, pending, officer, assigned)
 * @summary Get all dropdown options
 */
export const GetDropdownsResponse = zod.object({
  payroll: zod.array(zod.string()),
  company: zod.array(zod.string()),
  work: zod.array(zod.string()),
  pending: zod.array(zod.string()),
  officer: zod.array(zod.string()),
  assigned: zod.array(zod.string()),
});

/**
 * @summary Add a new option to a dropdown category
 */
export const AddDropdownOptionParams = zod.object({
  category: zod.enum([
    "payroll",
    "company",
    "work",
    "pending",
    "officer",
    "assigned",
  ]),
});

export const AddDropdownOptionBody = zod.object({
  value: zod.string(),
});

/**
 * @summary Remove an option from a dropdown category
 */
export const DeleteDropdownOptionParams = zod.object({
  category: zod.enum([
    "payroll",
    "company",
    "work",
    "pending",
    "officer",
    "assigned",
  ]),
});

export const DeleteDropdownOptionBody = zod.object({
  value: zod.string(),
});

/**
 * @summary List all form submissions
 */
export const listSubmissionsQueryLimitDefault = 50;
export const listSubmissionsQueryOffsetDefault = 0;

export const ListSubmissionsQueryParams = zod.object({
  limit: zod.coerce.number().default(listSubmissionsQueryLimitDefault),
  offset: zod.coerce.number().default(listSubmissionsQueryOffsetDefault),
});

export const ListSubmissionsResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      payroll: zod.string(),
      company: zod.string(),
      work: zod.string(),
      detail1: zod.string(),
      detail2: zod.string().nullable(),
      pending: zod.string(),
      remarks: zod.string().nullable(),
      officer: zod.string().nullable(),
      assigned: zod.string().nullable(),
      submittedAt: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Create a new form submission
 */
export const CreateSubmissionBody = zod.object({
  payroll: zod.string(),
  company: zod.string(),
  work: zod.string(),
  detail1: zod.string(),
  detail2: zod.string().nullish(),
  pending: zod.string(),
  remarks: zod.string().nullish(),
  officer: zod.string().nullish(),
  assigned: zod.string().nullish(),
});

/**
 * @summary Get a single submission
 */
export const GetSubmissionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSubmissionResponse = zod.object({
  id: zod.number(),
  payroll: zod.string(),
  company: zod.string(),
  work: zod.string(),
  detail1: zod.string(),
  detail2: zod.string().nullable(),
  pending: zod.string(),
  remarks: zod.string().nullable(),
  officer: zod.string().nullable(),
  assigned: zod.string().nullable(),
  submittedAt: zod.string(),
});

/**
 * @summary Delete a submission
 */
export const DeleteSubmissionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get submission statistics summary
 */
export const GetSubmissionSummaryQueryParams = zod.object({
  startDate: zod.coerce
    .string()
    .optional()
    .describe("Filter from this date (ISO format YYYY-MM-DD)"),
  endDate: zod.coerce
    .string()
    .optional()
    .describe("Filter up to this date (ISO format YYYY-MM-DD)"),
});

export const GetSubmissionSummaryResponse = zod.object({
  total: zod.number(),
  todayCount: zod.number(),
  byWork: zod.array(
    zod.object({
      work: zod.string(),
      count: zod.number(),
    }),
  ),
  byPending: zod.array(
    zod.object({
      pending: zod.string(),
      count: zod.number(),
    }),
  ),
  byOfficer: zod.array(
    zod.object({
      officer: zod.string(),
      total: zod.number(),
      done: zod.number(),
      pending: zod.number(),
    }),
  ),
  byDoer: zod.array(
    zod.object({
      doer: zod.string(),
      total: zod.number(),
      done: zod.number(),
      pending: zod.number(),
    }),
  ),
  recentSubmissions: zod.array(
    zod.object({
      id: zod.number(),
      payroll: zod.string(),
      company: zod.string(),
      work: zod.string(),
      detail1: zod.string(),
      detail2: zod.string().nullable(),
      pending: zod.string(),
      remarks: zod.string().nullable(),
      officer: zod.string().nullable(),
      assigned: zod.string().nullable(),
      submittedAt: zod.string(),
    }),
  ),
});