# lib/api-zod/src/generated/types/doerCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DoerCount {
  doer: string;
  total: number;
  done: number;
  pending: number;
}