# lib/api-zod/src/generated/types/dropdownData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DropdownData {
  payroll: string[];
  company: string[];
  work: string[];
  pending: string[];
  officer: string[];
  assigned: string[];
}