# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Submission {
  id: number;
  payroll: string;
  company: string;
  work: string;
  detail1: string;
  /** @nullable */
  detail2: string | null;
  pending: string;
  /** @nullable */
  remarks: string | null;
  /** @nullable */
  officer: string | null;
  /** @nullable */
  assigned: string | null;
  submittedAt: string;
}