# lib/api-zod/src/generated/types/submissionList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from "./submission";

export interface SubmissionList {
  items: Submission[];
  total: number;
}