# lib/api-zod/src/generated/types/submissionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DoerCount } from "./doerCount";
import type { OfficerCount } from "./officerCount";
import type { PendingCount } from "./pendingCount";
import type { Submission } from "./submission";
import type { WorkCount } from "./workCount";

export interface SubmissionSummary {
  total: number;
  todayCount: number;
  byWork: WorkCount[];
  byPending: PendingCount[];
  byOfficer: OfficerCount[];
  byDoer: DoerCount[];
  recentSubmissions: Submission[];
}